requirements.txt
flask
flask-cors
pymongo
python-chess
gunicorn
